requirements.txt
setuptools==68.2.2
plotly==5.18.0
dash==2.14.0
pandas==2.1.1
python-dotenv==1.0.0
these-united-states==1.1.0.21
numpy~=1.26.1